Split out `sim_relay` module
